Version 1

+++ src/main/resources/db/migration/V1.0__init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-  id int(20) NOT NULL AUTO_INCREMENT,
+  id bigint(20) NOT NULL AUTO_INCREMENT,
 username varchar(100) NOT NULL,
 first_name varchar(50) NOT NULL,
 last_name varchar(50) DEFAULT NULL,
